moved selector to an independent view
