+++ js/src/templates/compare/index.hbs
@@ -1,54 +1,23 @@
 <!-- compare banner -->
 
 <!-- country selectors  -->
-<div id="compare-countries-selectors" class="compare-countries-selectors">
+<div class="wrap">
+  <div id="selectors-container" class="compare-countries-selectors">
 
-  <div class="wrap">
 
-    <div class="selectors-container">
 
-      <div class="c-selector">
-        <div class="btn-drop-down">
-          <select  id="isoA" name="isoA">
-            <option value=""> select country </option>
-          </select>
-        </div>
-      </div>
+  </div>
 
-      <div class="c-selector">
-        <div class="btn-drop-down">
-          <select  id="isoB" name="isoB">
-            <option value=""> select country </option>
-          </select>
-        </div>
-      </div>
+  <!-- target cards  -->
+  <div class="target-list-container">
 
-      <div class="c-selector">
-        <div class="btn-drop-down">
-          <select  id="isoC" name="isoC">
-            <option value=""> select country </option>
-          </select>
-        </div>
-      </div>
+    <div class="wrap">
+
+      <div id="target-list" class="is-hidden"></div>
 
     </div>
 
-    <button class="c-button" type="button" id="compare-countries" name="compare-countries"> compare </button>
-
   </div>
-
 </div>
-
-<!-- target cards  -->
-<div class="target-list-container">
-
-  <div class="wrap">
-
-    <div id="target-list" class="is-hidden"></div>
-
-  </div>
-
-</div>
-
 
 <!-- map banner -->
