+++ js/src/templates/compare/selectors/selectors.hbs
@@ -1,0 +1,25 @@
+<div class="c-selector">
+  <div class="btn-drop-down">
+    <select  id="isoA" name="isoA">
+      <option value=""> select country </option>
+    </select>
+  </div>
+</div>
+
+<div class="c-selector">
+  <div class="btn-drop-down">
+    <select  id="isoB" name="isoB">
+      <option value=""> select country </option>
+    </select>
+  </div>
+</div>
+
+<div class="c-selector">
+  <div class="btn-drop-down">
+    <select  id="isoC" name="isoC">
+      <option value=""> select country </option>
+    </select>
+  </div>
+</div>
+
+<button class="c-button" type="button" id="compare-countries" name="compare-countries"> compare </button>
