+++ js/src/views/compare/compare.js
@@ -10,7 +10,8 @@ var TargetsCollection = require('../../collections/common/targets.js');
 
 var Status = require ('../../models/compare/status.js');
 
-var TargetListView = require('./targets/target_list.js');
+var SelectorsView = require('./selectors/selectors.js'),
+  TargetListView = require('./targets/target_list.js');
 
 var template = Handlebars.compile(require('../../templates/compare/index.hbs'));
 
@@ -31,26 +32,26 @@ var CompareView = Backbone.View.extend({
     this.status = new Status();
 
     // collections
-    this.countriesCollection = new CountriesCollection();
+    this.countriesCollection = CountriesCollection;
     this.targetsCollection = new TargetsCollection();
 
     // views
     this.targetListView;
+    this.selectorsView = new SelectorsView({
+      status: this.status
+    });
   },
 
   // this function initializes just before initial render
   show: function() {
     this._setView();
     this._setVars();
-    this._setListeners();
-
-    this._checkCompareButton();
 
     $.when(
       this.countriesCollection.getCountriesList(),
       this.targetsCollection.getTargetsList()
     ).done(function() {
-      this._populateSelectors();
+      this._renderSelectors();
       this._renderTargetList();
     }.bind(this));
   },
@@ -60,8 +61,6 @@ var CompareView = Backbone.View.extend({
   },
 
   _setVars: function() {
-    this.$selectors = this.$el.find('#compare-countries-selectors select');
-    this.$compareBtn = this.$el.find('#compare-countries');
     this.$targetList = this.$el.find('#target-list');
   },
 
@@ -79,88 +78,8 @@ var CompareView = Backbone.View.extend({
     });
   },
 
-  _setListeners: function() {
-    this.$selectors.on('change', this._setCountry.bind(this));
-    this.status.on('change', this._checkSelectors.bind(this));
-  },
-
-  _populateSelectors: function() {
-    var selectOptions;
-
-    this.countriesCollection.forEach(function(countryModel) {
-      var country = countryModel.toJSON();
-      selectOptions += '<option value="' + country.iso + '">' + country.name + '</option>';
-    });
-
-    this.$selectors.each(function(i, selector) {
-      $(selector).append(selectOptions);
-    });
-
-    var currentCountries = _.uniq(_.values(this.status.toJSON()));
-
-    if (!(_.compact(currentCountries).length)) {
-      return;
-    }
-
-    this.$selectors.each(function(i, selector) {
-      if (!currentCountries[i]) {
-        return;
-      }
-
-      this.value = currentCountries[i];
-    });
-
-    this._checkSelectors();
-  },
-
-  _checkSelectors: function() {
-    var newCountries = _.values(this.status.toJSON());
-
-    this.$selectors.each(function(i, selector) {
-
-      $(selector).find('option:disabled').removeAttr('disabled');
-
-      _.map(newCountries, function(iso, i) {
-
-        if (iso == '') {
-          return;
-        }
-
-        $(selector).find('option:not(:selected)[value="' + iso + '"]').attr('disabled', 'disabled');
-
-      }, this);
-
-    }.bind(this));
-  },
-
-  // set status with new incoming params
-  _setCountry: function(e) {
-    var selector = e.currentTarget;
-      id = selector.id,
-      iso = selector.value;
-
-    var newStatus = this.status.toJSON()
-    newStatus[id] = iso;
-
-    this.status.set(newStatus);
-
-    this._checkCompareButton();
-  },
-
-  // check if the compare button shoudl be available or not
-  _checkCompareButton: function() {
-    var settedCountries = _.compact(_.values(this.status.toJSON())),
-      totalCountries = settedCountries.length;
-
-    if (totalCountries > 1) {
-      if (this.$compareBtn.hasClass('-disabled')) {
-        this.$compareBtn.removeClass('-disabled');
-      }
-    } else {
-      if (!this.$compareBtn.hasClass('-disabled')) {
-        this.$compareBtn.addClass('-disabled');
-      }
-    }
+  _renderSelectors: function() {
+    this.$el.find('#selectors-container').html(this.selectorsView.render().el);
   },
 
   _compareCountries: function() {
